Minigrid with Conditional Distribution moved to tests/hierarchical_policies
